Moves health check components into separate health package
- Adds comment to Dockerfile
- Adds health check property to databridge in dockercompose
