+++ basyx.components.databridge.core/pom.xml
@@ -32,11 +32,6 @@
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-surefire-plugin</artifactId>
 				<version>3.0.0-M8</version>
-				<configuration>
-					<excludes>
-						<exclude>**/MongoPopulatorTool.java</exclude>
-					</excludes>
-				</configuration>
 			</plugin>
 		</plugins>
 	</build>
