+++ basyx.components.databridge/docker-compose.yml
@@ -12,6 +12,12 @@ services:
         condition: service_healthy
       aas:
         condition: service_healthy
+    healthcheck:
+      test: wget --no-verbose --tries=1 databridge:8085/health || exit 1
+      interval: 20s
+      retries: 3
+      start_period: 10s
+      timeout: 10s
 
   aas:
     build:
